JS and angular optimizations done
Js concatenation and minification done, angular partials are also put in
templatecache to get rid of all requests for pages. Sass however still
doesn't work inside vs, seems to be a problem with web essentials, tried
bundletransformer and it worked, but it was really slow. Sticking with
compiling it outside of vs for now.